d1: Enable building LRADC driver for the linux kernel
The D1 has a low-rate analog-to-digital converter that performs
low resolution, low-rate sampling. It is intended to be used with a
set of buttons, each supplying a different voltage. This allows for
chaining many buttons on a single pin of the CPU.

Some of the dev boards  supported by OpenWRT have buttons
attached to this.

Signed-off-by: Gabriel Roper <lockheedmaniac@gmail.com>